Minor changes to avoid compiler warnings
git-svn-id: https://subversion.xor.aps.anl.gov/synApps/areaDetector/trunk@15509 dc6c5ff5-0b8b-c028-a01f-ffb33f00fc8b

+++ ADApp/roperSrc/roper.cpp
@@ -153,7 +153,7 @@ private:
 };
 
 
-#define NUM_ROPER_PARAMS (&LAST_ROPER_PARAM - &FIRST_ROPER_PARAM + 1)
+#define NUM_ROPER_PARAMS ((int)(&LAST_ROPER_PARAM - &FIRST_ROPER_PARAM + 1))
 
 /* Convert from a C string to a BSTR.  This must be freed by the caller! */
 BSTR stringToBSTR(char *str)
